+++ uni_modules/uni-badge/changelog.md
@@ -1,3 +1,9 @@
+## 1.1.0（2021-05-12）
+- 新增 uni-badge 的 absolute 属性，支持定位
+- 新增 uni-badge 的 offset 属性，支持定位偏移
+- 新增 uni-badge 的 is-dot 属性，支持仅显示有一个小点
+- 新增 uni-badge 的 max-num 属性，支持自定义封顶的数字值，超过 99 显示99+
+- 优化 uni-badge 属性 custom-style， 支持以对象形式自定义样式
 ## 1.0.7（2021-05-07）
 - 修复 uni-badge 在 App 端，数字小于10时不是圆形的bug
 - 修复 uni-badge 在父元素不是 flex 布局时，宽度缩小的bug